extracted schema validation to reuse in echo-handler
