+++ packages/core/echo/echo-schema/src/effect/typed-handler.ts
@@ -2,16 +2,10 @@
 // Copyright 2024 DXOS.org
 //
 
-import * as AST from '@effect/schema/AST';
-import * as S from '@effect/schema/Schema';
-
 import { compositeRuntime } from '@dxos/echo-signals/runtime';
 
 import { type ReactiveHandler, createReactiveProxy, isValidProxyTarget } from './proxy';
-import { ReactiveArray } from './reactive-array';
-
-export const symbolSchema = Symbol.for('@dxos/schema');
-export const symbolTypeAst = Symbol.for('@dxos/type/AST');
+import { SchemaValidator } from './schema-validator';
 
 export class TypedReactiveHandler<T extends object> implements ReactiveHandler<T> {
   _proxyMap = new WeakMap<object, any>();
@@ -19,17 +13,7 @@ export class TypedReactiveHandler<T extends object> implements ReactiveHandler<T
   _isInSet = false;
 
   _init(target: any): void {
-    this.makeArraysReactive(target);
-  }
-
-  private makeArraysReactive(target: any) {
-    for (const key in target) {
-      if (Array.isArray(target[key]) && !(target[key] instanceof ReactiveArray)) {
-        target[key] = ReactiveArray.from(target[key]);
-        const schema = this.getPropertySchema(target, key);
-        setSchemaProperties(target[key], schema);
-      }
-    }
+    SchemaValidator.initTypedTarget(target);
   }
 
   get(target: any, prop: string | symbol, receiver: any): any {
@@ -45,7 +29,7 @@ export class TypedReactiveHandler<T extends object> implements ReactiveHandler<T
   set(target: any, prop: string | symbol, value: any, receiver: any): boolean {
     try {
       this._isInSet = true;
-      const validatedValue = this.validateValue(target, prop, value);
+      const validatedValue = SchemaValidator.validateValue(target, prop, value);
       const result = Reflect.set(target, prop, validatedValue, receiver);
       this._signal.notifyWrite();
       return result;
@@ -58,7 +42,7 @@ export class TypedReactiveHandler<T extends object> implements ReactiveHandler<T
     if (typeof property === 'symbol' || this._isInSet) {
       return Reflect.defineProperty(target, property, attributes);
     }
-    const validatedValue = this.validateValue(target, property, attributes.value);
+    const validatedValue = SchemaValidator.validateValue(target, property, attributes.value);
     const result = Reflect.defineProperty(target, property, {
       ...attributes,
       value: validatedValue,
@@ -66,62 +50,4 @@ export class TypedReactiveHandler<T extends object> implements ReactiveHandler<T
     this._signal.notifyWrite();
     return result;
   }
-
-  private validateValue(target: any, prop: string | symbol, value: any) {
-    const schema = this.getPropertySchema(target, prop);
-    const _ = S.asserts(schema)(value);
-    if (Array.isArray(value)) {
-      value = new ReactiveArray(...value);
-    }
-    if (isValidProxyTarget(value)) {
-      setSchemaProperties(value, schema);
-    }
-    return value;
-  }
-
-  private getPropertySchema(target: any, prop: string | symbol): S.Schema<any> {
-    if (target instanceof ReactiveArray) {
-      return getArrayElementSchema((target as any)[symbolSchema]);
-    }
-    const ast = target[symbolTypeAst] as AST.AST;
-    const properties = AST.getPropertySignatures(ast).find((property) => property.name === prop);
-    if (!properties) {
-      throw new Error(`Invalid property: ${prop.toString()}`);
-    }
-    return S.make(properties.type);
-  }
 }
-
-const getArrayElementSchema = (arraySchema: S.Schema<any>): S.Schema<any> => {
-  return S.make((arraySchema.ast as any).rest.value[0]);
-};
-
-/**
- * Recursively set AST property on the object.
- */
-// TODO(burdon): Use visitProperties.
-export const setSchemaProperties = (obj: any, schema: S.Schema<any>) => {
-  if (!obj[symbolSchema]) {
-    Object.defineProperty(obj, symbolSchema, { enumerable: false, value: schema });
-  }
-  if (AST.isTypeLiteral(schema.ast)) {
-    Object.defineProperty(obj, symbolTypeAst, {
-      enumerable: false,
-      value: schema.ast,
-    });
-
-    for (const property of schema.ast.propertySignatures) {
-      const value = (obj as any)[property.name];
-      if (isValidProxyTarget(value)) {
-        setSchemaProperties(value, S.make(property.type));
-      }
-    }
-  } else if (Array.isArray(obj) && AST.isTuple(schema.ast)) {
-    const elementSchema = getArrayElementSchema(schema);
-    for (const value of obj) {
-      if (isValidProxyTarget(value)) {
-        setSchemaProperties(value, elementSchema);
-      }
-    }
-  }
-};
